InternShipCandidat fonctionnel

+++ src/main/java/com/sap/ose/projetose/dto/InternOfferDto.java
@@ -24,7 +24,7 @@ public class InternOfferDto {
     private String startDate;
     private String endDate;
     private List<InternshipCandidatesDto> internshipCandidates;
-    private int programmeId;
+    private long programmeId;
     private FileDto file;
     private long employeurId;
     private String programmeNom;
@@ -52,7 +52,7 @@ public class InternOfferDto {
         this.state = internOffer.getState();
     }
 
-    public InternOfferDto(String title, String location, String description, double salaryByHour, String startDate, String endDate, List<InternshipCandidatesDto> internshipCandidates, int programmeId, FileDto file,State state) {
+    public InternOfferDto(String title, String location, String description, double salaryByHour, String startDate, String endDate, List<InternshipCandidatesDto> internshipCandidates, long programmeId, FileDto file,State state) {
         this.title = title;
         this.location = location;
         this.description = description;
