+++ src/main/java/com/sap/ose/projetose/repository/ProgrammeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProgrammeRepository extends JpaRepository<Programme, Integer> {
+public interface ProgrammeRepository extends JpaRepository<Programme, Long> {
 }
